Extracted parser primary exp tests to own file
