+++ spec/Parser/Parser_spec.js
@@ -28,70 +28,6 @@ describe('Parser', () => {
     });
   });
 
-  describe('should handle primary expressions', () => {
-    it('NUMBER', () => {
-      const tokens = [
-        new Token(TokenType.NUMBER, '9', 9, 1),
-        SEMICOLON_TOKEN,
-        EOF_TOKEN
-      ];
-
-      expect(astPrinter.print(parser.parse(tokens))).toEqual('( statement 9 )');
-    });
-
-    it('STRING', () => {
-      const tokens = [
-        new Token(TokenType.STRING, '\'kthnxbai\'', 'kthnxbai', 1),
-        SEMICOLON_TOKEN,
-        EOF_TOKEN
-      ];
-
-      expect(astPrinter.print(parser.parse(tokens))).toEqual('( statement kthnxbai )');
-    });
-
-    it('false', () => {
-      const tokens = [
-        new Token(TokenType.FALSE, 'false', undefined, 1),
-        SEMICOLON_TOKEN,
-        EOF_TOKEN
-      ];
-
-      expect(astPrinter.print(parser.parse(tokens))).toEqual('( statement false )');
-    });
-
-    it('true', () => {
-      const tokens = [
-        new Token(TokenType.TRUE, 'true', undefined, 1),
-        SEMICOLON_TOKEN,
-        EOF_TOKEN
-      ];
-
-      expect(astPrinter.print(parser.parse(tokens))).toEqual('( statement true )');
-    });
-
-    it('nil', () => {
-      const tokens = [
-        new Token(TokenType.NIL, 'nil', undefined, 1),
-        SEMICOLON_TOKEN,
-        EOF_TOKEN
-      ];
-
-      expect(astPrinter.print(parser.parse(tokens))).toEqual('( statement null )');
-    });
-
-    it('( EXPRESSION )', () => {
-      const tokens = [
-        new Token(TokenType.LEFT_PAREN, '(', undefined, 1),
-        new Token(TokenType.TRUE, 'true', undefined, 1),
-        new Token(TokenType.RIGHT_PAREN, ')', undefined, 1),
-        SEMICOLON_TOKEN,
-        EOF_TOKEN
-      ];
-
-      expect(astPrinter.print(parser.parse(tokens))).toEqual('( statement ( group true ) )');
-    });
-  });
-
   describe('should handle unary expressions', () => {
     it('!true', () => {
       const tokens = [
